+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Major rewrite of internals. Should be much easier to implement new features
 * Removed `featureName` from `jefe finish`
+* Allow project names to differ from repo names
+* New command to generate Intellij vcs.xml file
 
 ## 0.2.10